move all api calls to app.service
