--- electron/ui/src/services/homepage.service.js
@@ -1,8 +1,0 @@
-
-export const runModel = (data) => {
-    return fetch('http://localhost:8001/run_model/', {
-        method: 'POST', 
-        mode: 'cors',
-        body: JSON.stringify(data)
-    });
-};
