--- electron/ui/src/services/sidebar.service.js
@@ -1,10 +1,0 @@
-export const uploadExcelSheet = (data, name) => {
-    return fetch('http://localhost:8001/upload/'+name, {
-        method: 'POST', 
-        mode: 'cors',
-        body: data
-    });
-}; 
-
-
-
